+++ lib/supabaseServer.ts
@@ -1,0 +1,13 @@
+import { cookies, headers } from 'next/headers';
+import { createServerClient } from '@supabase/ssr';
+import { Database } from '@/lib/database.types';
+
+export const supabaseServer = () =>
+  createServerClient<Database>(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      cookies,
+      headers,
+    }
+  );